tests/requirements-tests.txt
pytest 
pytest-asyncio
pytest-aiohttp
playwright 
pytest-playwright
flake8
flake8-pyproject
black
pytest-cov